+++ app/views/user/mobil/table_view.php
@@ -1,15 +1,18 @@
 <div class="container shadow p-4 mb-5 bg-white rounded">
     <div class="d-flex justify-content-between">
-        <form class="d-none d-sm-inline-block form-inline mr-auto ml-md-3 my-2 my-md-0 mw-100 navbar-search">
+    <form class="d-none d-sm-inline-block form-inline mr-auto ml-md-3 my-2 my-md-0 mw-100 navbar-search" method="post" action="<?=BASEURL?>/user/cari_mobil">
             <div class="input-group">
-                <input type="text" class="form-control bg-light border-1 small" placeholder="Search for..." aria-label="Search" aria-describedby="basic-addon2">
+                <input type="text" name="keyword" class="form-control bg-light border-1 small" placeholder="Search for..." aria-label="Search" aria-describedby="basic-addon2" autocomplete="off">
                 <div class="input-group-append">
-                <button class="btn btn-primary" type="button">
+                <button class="btn btn-primary" type="submit">
                     <i class="fas fa-search fa-sm"></i>
                 </button>
                 </div>
             </div>
         </form>
+        <div>
+            <a class="btn btn-success second-view" href="<?= BASEURL?>/user/mobil">Table View</a>
+        </div>
         <div>
         <button type="button" name="" id="" class="btn btn-primary btn-icon-split" data-toggle="modal" data-target=".modalTambahMobil">
             <span class="icon text-white-50">
@@ -21,7 +24,7 @@
     </div>
     <div class="row ml-5">
         <?php foreach($data['mobil'] as $mobil)   :?>
-            <div class="col-lg-6 card-deck mt-3">
+            <div class="col-lg-4 card-deck mt-3">
                 <div class="card shadow-lg p-3 mb-5 bg-white rounded" style="width: 18rem;">
                     <div class="card-header bg-white" style="font-size:1.4em;"><?= $mobil['nama_mobil']?></div>
                     <div class="card-body">
@@ -45,7 +48,7 @@
                     </div>
                     <div class="card-footer bg-white">
                         <a href="<?= BASEURL?>/user/detail_mobil/<?= $mobil['id_mobil']?>" class="btn btn-success float-right">Detail</a>
-                        <a href="<?= BASEURL?>/user/hapus_mobil/<?= $mobil['id_mobil']?>" class="btn btn-danger float-right mr-2">Hapus</a>
+                        <a href="<?= BASEURL?>/user/hapus_mobil/<?= $mobil['id_mobil']?>/<?= $data['function']?>" onclick="return confirm('Yakin Untuk Menghapus Mobil Ini ?');" class="btn btn-danger float-right mr-2">Hapus</a>
                     </div>
                 </div>
             </div>
